Move useUsingKeyboardContext test to UsingKeyboardProvider.spex.tsx
